% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_num_id_col.R
\name{add_num_id_col}
\alias{add_num_id_col}
\title{Adds a column with numeric IDs}
\usage{
add_num_id_col(long_data)
}
\arguments{
\item{long_data}{A data.frame with at least column \code{ID}.}
}
\value{
A data.frame with at least columns \code{ID}, and \code{ID_num}.
}
\description{
Adds a new column to the data.frame with numeric IDs which is
called \code{ID_num}.
}
\examples{
long_data <- tibble::tibble(ID = c("KT011", "KT002", "KT011"), 
                            something = c("t1", "t2", "t3"))
add_num_id_col(long_data)

}
\author{
Kira E. Detrois
}
