% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_df_multi_icd_ver.R
\name{create_test_df_multi_icd_ver}
\alias{create_test_df_multi_icd_ver}
\title{Creates a data.frame with random ICD-code entries}
\usage{
create_test_df_multi_icd_ver(
  n_icd10 = 50,
  n_icd9 = 20,
  icd10_indv = c("KT0000001", "KT0000002", "KT0000004", "KT0000005", "KT0000006"),
  icd9_indv = c("KT0000005", "KT0000007", "KT0000002", "KT0000008", "KT0000009")
)
}
\arguments{
\item{n_icd10}{Number of draws from ICD10}

\item{n_icd9}{Number of draws from ICD9}

\item{icd10_indv}{A character vector.
The names of the individuals with ICD10 codes.}

\item{icd9_indv}{A character vector.
The names of the individuals with ICD9 codes.}
}
\value{
A tibble with columns \code{ID}, \code{Event_age}, \code{primary_ICD},
and \code{secondary_ICD}. Filled with random ICD codes, and age at event.
}
\description{
Creates a tibble with random ICD-codes entries from different
ICD-versions using the function \code{\link{create_test_df}}.
The results are meant to mimic INTERVENE longitudinal files for
testing.
}
\details{
It is possible to create a mix of ICD-10 and ICD-9 codes to create
a more realistic testing dataset.
}
\author{
Kira E. Detrois
}
