% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comorb_icd_ver_str.R
\name{get_comorb_icd_ver_str}
\alias{get_comorb_icd_ver_str}
\title{Translates ICD-version descriptions}
\usage{
get_comorb_icd_ver_str(icd_version)
}
\arguments{
\item{icd_version}{The ICD-version in INTERVENE format.
Can be either "10", "10CM", "9", or "9CM".}
}
\value{
A string. Either "charlson_icd10_quan", or
"charlson_icd9_quan".
}
\description{
Translates the ICD-version descriptions "10", "10CM", "9", and "9CM"
from an INTERVENE longitudinal file into strings understandable
by the function \code{\link[comorbidity]{comorbidity}}.
}
\author{
Kira E. Detrois
}
