% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_long_data.R
\name{get_group_long_data}
\alias{get_group_long_data}
\title{Gets the data for the current ICD-version}
\usage{
get_group_long_data(group_long_data, icd_version)
}
\arguments{
\item{group_long_data}{A list. This should be created, using
function
\code{\link{group_long_data_by_icd_ver}}.}

\item{icd_version}{The ICD-version in INTERVENE format.
This should be "10", "10CM", "9", or "9CM".}
}
\value{
The data.frame with data for the chose ICD-version.
}
\description{
Gets the entries of the data.frame for the selected ICD-version.
}
\examples{
long_data <- tibble::tibble(ICD_version=c("10", "9", "10", "10CM"),
                            primary_ICD=c("t1", "t2", "t3", "t4"))
group_long_data <- group_long_data_by_icd_ver(long_data)
get_group_long_data(group_long_data, "10")

}
\author{
Kira E. Detrois
}
