% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_long_data_by_icd_ver.R
\name{group_long_data_by_icd_ver}
\alias{group_long_data_by_icd_ver}
\title{Groups the data according to the ICD-versions}
\usage{
group_long_data_by_icd_ver(long_data)
}
\arguments{
\item{long_data}{A data.frame with at least column \code{ICD_version}.}
}
\value{
A list(\code{groups}, \code{group_keys}, \code{group_idxs}):
\itemize{
\item \code{groups}: The grouped tibble.
\item \code{group_keys}: A list. The name of each group, so the
ICD-versions present in the data.
\item \code{group_idxs}: A named list. The indices of the
elements of each group in the original
\code{long_data}. The names are the group
names.
}
}
\description{
Groups the data.frame according to the ICD-versions, using
function \code{\link[dplyr]{group_by}}.
}
