% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cci_scores_to_num_ids.R
\name{map_cci_scores_to_num_ids}
\alias{map_cci_scores_to_num_ids}
\title{Map CCI scores onto IDs}
\usage{
map_cci_scores_to_num_ids(cci_scores, IDs_num)
}
\arguments{
\item{cci_scores}{A data.frame with at least column \code{ID_num}, and
column \code{score}.}

\item{IDs_num}{A numeric. The numeric IDs of the individuals.}
}
\value{
A tibble with columns \code{ID_num}, and \code{score}.
}
\description{
Maps the scores obtained by \code{\link[comorbidity]{score}} back
onto the individuals numeric IDs
}
\details{
CAREFUL!! The function score sorts the individuals by their
(numeric) names.
}
\author{
Kira E. Detrois
}
